@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
+    compileSdkVersion 29    
     buildToolsVersion "23.0.1"
 
     defaultConfig {
         applicationId "com.dlka.psonememorymanager"
         minSdkVersion 16
-        targetSdkVersion 23
+        targetSdkVersion 29
         versionName '0.5 free'
         manifestPlaceholders = [ applicationName:"PSOne-MemoryCard-Manager" ]
         manifestPlaceholders = [ applicationCode:"10000000" ]
@@ -28,7 +28,7 @@ android {
 dependencies {
     compile 'com.google.android.gms:play-services-ads:8.1.0'
     compile 'com.google.android.gms:play-services-analytics:8.1.0'
-    compile 'com.android.support:support-v4:23.1.0'
+    compile 'androidx.legacy:legacy-support-v4:1.0.0'
     compile 'cn.pedant.sweetalert:library:1.3'
     compile 'fr.nicolaspomepuy:discreetapprate:2.0.3@aar'
     compile 'org.sufficientlysecure:donations:2.3'
